% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{bind_rows}
\alias{bind_rows}
\alias{bind_cols}
\alias{inner_join}
\alias{right_join}
\title{Inner join datasets}
\usage{
bind_rows(..., .id = NULL, add.cell.ids = NULL)

bind_cols(..., .id = NULL)
}
\arguments{
\item{...}{Data frames to combine (See dplyr)}

\item{.id}{Data frame identifier.

When \code{.id} is supplied, a new column of identifiers is
created to link each row to its original data frame. The labels
are taken from the named arguments to \code{bind_rows()}. When a
list of data frames is supplied, the labels are taken from the
names of the list. If no names are found a numeric sequence is
used instead.}

\item{add.cell.ids}{from SpatialExperiment 3.0 A character vector of
length(x=c(x, y)). Appends the corresponding values to the start of each
objects' cell names.}

\item{x}{tbls to join. (See dplyr)}

\item{y}{tbls to join. (See dplyr)}

\item{by}{A character vector of variables to join by. (See dplyr)}

\item{copy}{If x and y are not from the same data source, and copy is TRUE,
then y will be copied into the same src as x. (See dplyr)}

\item{suffix}{If there are non-joined duplicate variables in x and y, these
suffixes will be added to the output to disambiguate them. Should be a
character vector of length 2. (See dplyr)}
}
\value{
A tidySpatialExperiment object

A tidySpatialExperiment object
}
\description{
Inner join datasets

Right join datasets
}
\examples{
`\%>\%` <- magrittr::`\%>\%`
example(read10xVisium)

spe \%>\%
    inner_join(
        spe \%>\%
            filter(in_tissue == TRUE) \%>\%
            mutate(new_column = 1)
        )

`\%>\%` <- magrittr::`\%>\%`
example(read10xVisium)

spe \%>\%
    right_join(
        spe \%>\%
            filter(in_tissue == TRUE) \%>\%
            mutate(new_column = 1)
        )

}
