% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ellipse}
\alias{ellipse}
\title{Ellipse Gating Function}
\usage{
ellipse(spatial_coord1, spatial_coord2, center, axes_lengths)
}
\arguments{
\item{spatial_coord1}{Numeric vector for x-coordinates}

\item{spatial_coord2}{Numeric vector for y-coordinates}

\item{center}{Numeric vector (length 2) for ellipse center (x, y)}

\item{axes_lengths}{Numeric vector (length 2) for the lengths of the major and minor axes of the ellipse}
}
\value{
Logical vector indicating points within the ellipse
}
\description{
Function to create an ellipse gate in a SpatialExperiment object
}
\examples{
example(read10xVisium)
spe |>
  mutate(in_ellipse = ellipse(array_col, array_row, center = c(50, 50), axes_lengths = c(20, 10)))
}