% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{join_features}
\alias{join_features}
\title{Extract and join information for features.}
\arguments{
\item{.data}{A SpatialExperiment object}

\item{features}{A vector of feature identifiers to join}

\item{all}{If TRUE return all}

\item{exclude_zeros}{If TRUE exclude zero values}

\item{shape}{Format of the returned table "long" or "wide"}

\item{...}{Parameters to pass to join wide, i.e. assay name to extract feature abundance from and gene prefix, for shape="wide"}
}
\value{
An object containing the information.for the specified features
}
\description{
join_features() extracts and joins information for specified features
}
\details{
This function extracts information for specified features and returns the information in either long or wide format.
}
\examples{
example(read10xVisium)
spe |>
    join_features(features = "ENSMUSG00000025900")
}
