% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr_methods.R
\name{nest}
\alias{nest}
\title{nest}
\arguments{
\item{.data}{A tbl. (See tidyr)}

\item{...}{Name-variable pairs of the form new_col=c(col1, col2, col3) (See tidyr)}

\item{.names_sep}{See ?tidyr::nest}
}
\value{
A tidySpatialExperiment objector a tibble depending on input
}
\description{
nest
}
\examples{

`\%>\%` <- magrittr::`\%>\%`
example(read10xVisium)

spe \%>\%
    nest(data = -sample_id)

}
