% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{pull}
\alias{pull}
\title{Extract a single column}
\value{
A vector the same size as \code{.data}.
}
\description{
\code{pull()} is similar to \code{$}. It's mostly useful because it looks a little
nicer in pipes, it also works with remote data frames, and it can optionally
name the output.
}
\section{Methods}{


This function is a \strong{generic}, which means that packages can provide
implementations (methods) for other classes. See the documentation of
individual methods for extra arguments and differences in behaviour.

The following methods are currently available in loaded packages:
\Sexpr[stage=render,results=rd]{dplyr:::methods_rd("pull")}.

}

\examples{
example(read10xVisium)
spe |>
    pull(in_tissue)

}
