% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rectangle}
\alias{rectangle}
\title{Rectangle Gating Function}
\usage{
rectangle(spatial_coord1, spatial_coord2, center, height, width)
}
\arguments{
\item{spatial_coord1}{Numeric vector for x-coordinates (e.g., array_col)}

\item{spatial_coord2}{Numeric vector for y-coordinates (e.g., array_row)}

\item{center}{Numeric vector of length 2 specifying the center of the rectangle (x, y)}

\item{height}{The height of the rectangle}

\item{width}{The width of the rectangle}
}
\value{
Logical vector indicating points within the rectangle
}
\description{
Determines whether points specified by spatial coordinates are within a defined rectangle.
}
\examples{
example(read10xVisium)
spe |>
    mutate(in_rectangle = rectangle(array_col, array_row, center = c(50, 50), height = 20, width = 10))

}