% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy}
\alias{tidy}
\title{tidy for SpatialExperiment}
\usage{
tidy(object)
}
\arguments{
\item{object}{A SpatialExperiment object}
}
\value{
A tidySpatialExperiment object
}
\description{
tidy for SpatialExperiment
}
\examples{

tidySpatialExperiment::pbmc_small
}
