% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{bind_cols}
\alias{bind_cols}
\title{Efficiently bind multiple data frames by row and column}
\value{
`bind_rows()` and `bind_cols()` return the same type as
  the first input, either a data frame, `tbl_df`, or `grouped_df`.

`bind_rows()` and `bind_cols()` return the same type as
  the first input, either a data frame, `tbl_df`, or `grouped_df`.
}
\description{
This is an efficient implementation of the common pattern of
`do.call(rbind, dfs)` or `do.call(cbind, dfs)` for binding many
data frames into one.

This is an efficient implementation of the common pattern of
`do.call(rbind, dfs)` or `do.call(cbind, dfs)` for binding many
data frames into one.
}
\details{
The output of `bind_rows()` will contain a column if that column
appears in any of the inputs.

The output of `bind_rows()` will contain a column if that column
appears in any of the inputs.
}
\examples{
# Note: "dplyr" does not provide a generic function for `bind_cols`. Therefore, the generic 
# function `bind_cols` located in "ttservice" should be called explicitly with 
# `ttservice::bind_cols` to avoid conflicts.

example(read10xVisium)
spe |>
    ttservice::bind_cols(1:99)
}
