% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gate_interactive}
\alias{gate_interactive}
\title{Gate interactive}
\usage{
gate_interactive(
  spe,
  image_index = 1,
  colour = NULL,
  shape = NULL,
  alpha = 1,
  size = 2
)
}
\arguments{
\item{spe}{A SpatialExperiment object}

\item{image_index}{The image to display if multiple are stored within the provided
SpatialExperiment object.}

\item{colour}{A string representing the point colour column. Or, a colour code string compatible
with ggplot2.}

\item{shape}{A string representing the point shape column, coercible to a factor. Or, a shape code
numeric compatible with ggplot2 (0-127)}

\item{alpha}{A string representing the point alpha column, coercible to a factor. Or, an alpha numeric
compatible with ggplot2 (0-1).}

\item{size}{A string representing the point shape column, coercible to a factor. Or, a size numeric
compatible with ggplot2 (0-20).}
}
\value{
The input SpatialExperiment object with a new column \code{.gate_interactive}, recording the
gates each X and Y coordinate pair is within. A record of the selected points is stored in
\code{tidygate_env$select_data} and a record of the gates is stored in \code{tidygate_env$brush_data}.
}
\description{
Interactively gate points by their location in space, with image data overlaid.
}
\examples{
\dontrun{
example(read10xVisium)
spe |>
    gate_interactive()
}
}
