% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gate_interactive}
\alias{gate_interactive}
\title{Gate interactive}
\usage{
gate_interactive(
  spe,
  image_index = 1,
  colour_column = NULL,
  shape_column = NULL,
  alpha = 1,
  size = 2
)
}
\arguments{
\item{spe}{A SpatialExperiment object}

\item{image_index}{The image to display if multiple are stored within the provided
SpatialExperiment object.}

\item{colour_column}{A column name string representing the point colour.}

\item{shape_column}{A column name string representing the point shape. Must be coercible to
a factor.}

\item{alpha}{A numeric value representing the opacity of points, with 1 being completely opaque
and 0 being completely transparent.}

\item{size}{A numeric value representing the size of points.}
}
\value{
A vector of lists, recording the gates each X and Y coordinate pair is within. A record
of the selected points is stored in \code{tidygate_env$select_data} and a record of the gates is
stored in \code{tidygate_env$brush_data}.

The input SpatialExperiment object with a new column \code{.gate_interactive}, recording the
gates each X and Y coordinate pair is within. A record of the selected points is stored in
\code{tidygate_env$select_data} and a record of the gates is stored in \code{tidygate_env$brush_data}.
}
\description{
Interactively gate points by their location in space, with image data overlaid.
}
\examples{
\dontrun{
example(read10xVisium)
spe |>
    gate_interactive()
}
}
